Reviews with higher stars are useless
Preliminary exploration of the reviews table revealed that there is a negative
correlation between the stars and useful votes of a review. A possible
explaination for this would be that the stars received by a review is tampered
by fake reviewers however real users don't rate these reviews useful.

UPDATE: turns out the stars on a review is given by the reviewer in context of
the business, so the behaviour observered above is expected as the rating given
to a business by a reviewer should not be related to the useful votes received
by that review.