Test run with Naive Bayes
I wanted to do a full iteration ie. compare 2 model accuracy to determine
existence/non existence of bias whilst rating usefulness of a review. X was not
preprocessed, neither was it normalized. I used the Naive Bayes classifier since
it is simple and runs quickly, the hyper parameters were not changed.
CountVectorize was used for feature extraction of the text.

I was able to train the classifier with just the review text and got an accuracy
of 61%. The problem was training model 2 with both the text and the useful
votes. This was challenging since the datatypes are mixed and CountVectorize
expects a list of strings (it cannot handle higher dim dfs). Two approaches were
tried:

1. Convert useful votes column to a sparse matrix (using
scipy.sparse.csc_matrix) and then stack it (scipy.sparse.hstack).

2. Read the reviews df as string entirely (ie. both column as string) and then
vectorize both columns and then stack.

Both methods produce errors, also scipy has multiple ways of generating sparse
matrices and stacking so it gets more confusing.