Length of reviews as feature
Since bag of words wasn't working, I tried to simply take the length of each
review as a feature. Got 52% for model 1 and 99% for model 2!

The numbers are bogus and will definitely go done once we do preprocesssing and
better feature extraction but for now this is a good proof of concept!